# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Universal SOP Architect",
    "description": "专家经验数字化编译平台 - 将业务专家的直觉与流程编译为 AI 可执行的协议",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/workflows": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "List Workflows",
        "description": "List all workflows with pagination.\n\n- **page**: Page number (default: 1)\n- **page_size**: Number of items per page (default: 20, max: 100)",
        "operationId": "list_workflows_api_workflows_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Create Workflow",
        "description": "Create a new workflow.\n\n- **name**: Workflow name (required)\n- **description**: Workflow description (optional)\n- **cover_image_url**: Cover image URL (optional)\n- **status**: Workflow status, 'draft' or 'deployed' (default: 'draft')",
        "operationId": "create_workflow_api_workflows_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workflows/{workflow_id}/steps": {
      "get": {
        "tags": [
          "steps"
        ],
        "summary": "List Steps",
        "description": "List all steps for a workflow.\n\n- **workflow_id**: UUID of the workflow\n- **page**: Page number (default: 1)\n- **page_size**: Number of items per page (default: 100, max: 100)",
        "operationId": "list_steps_api_workflows__workflow_id__steps_get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 100,
              "title": "Page Size"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkflowStepResponse"
                  },
                  "title": "Response List Steps Api Workflows  Workflow Id  Steps Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "steps"
        ],
        "summary": "Create Step",
        "description": "Create a new step in a workflow.\n\n- **workflow_id**: UUID of the workflow\n- **auto_order**: If true, automatically assign step_order (append to end)\n- **name**: Step name (required)\n- **step_order**: Step order (required if auto_order is false)\n- **status**: Step status, 'pending' or 'completed' (default: 'pending')",
        "operationId": "create_step_api_workflows__workflow_id__steps_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Id"
            }
          },
          {
            "name": "auto_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Auto-assign step order (append to end)",
              "default": false,
              "title": "Auto Order"
            },
            "description": "Auto-assign step order (append to end)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowStepCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStepResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workflows/{workflow_id}/steps/{step_id}": {
      "get": {
        "tags": [
          "steps"
        ],
        "summary": "Get Step",
        "description": "Get a step by ID.\n\n- **workflow_id**: UUID of the workflow\n- **step_id**: UUID of the step",
        "operationId": "get_step_api_workflows__workflow_id__steps__step_id__get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Id"
            }
          },
          {
            "name": "step_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Step Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStepResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "steps"
        ],
        "summary": "Update Step",
        "description": "Update a step.\n\n- **workflow_id**: UUID of the workflow\n- **step_id**: UUID of the step\n- All fields are optional for partial updates",
        "operationId": "update_step_api_workflows__workflow_id__steps__step_id__put",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Id"
            }
          },
          {
            "name": "step_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Step Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowStepUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStepResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "steps"
        ],
        "summary": "Delete Step",
        "description": "Delete a step.\n\n- **workflow_id**: UUID of the workflow\n- **step_id**: UUID of the step\n\nNote: Remaining steps will be automatically reordered.",
        "operationId": "delete_step_api_workflows__workflow_id__steps__step_id__delete",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Id"
            }
          },
          {
            "name": "step_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Step Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workflows/{workflow_id}/steps/{step_id}/branches": {
      "post": {
        "tags": [
          "steps"
        ],
        "summary": "Add Routing Branch",
        "description": "Add a routing branch to a step.\n\n- **workflow_id**: UUID of the workflow\n- **step_id**: UUID of the step\n- **condition_result**: Condition result (e.g., 'FAIL', 'UNSTABLE')\n- **action_type**: Action type (e.g., 'REJECT', 'ESCALATE')\n- **next_step_id**: Next step ID or 'end_process'",
        "operationId": "add_routing_branch_api_workflows__workflow_id__steps__step_id__branches_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Id"
            }
          },
          {
            "name": "step_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Step Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutingBranchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStepResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workflows/{workflow_id}/steps/{step_id}/branches/{branch_id}": {
      "delete": {
        "tags": [
          "steps"
        ],
        "summary": "Remove Routing Branch",
        "description": "Remove a routing branch from a step.\n\n- **workflow_id**: UUID of the workflow\n- **step_id**: UUID of the step\n- **branch_id**: UUID of the branch to remove",
        "operationId": "remove_routing_branch_api_workflows__workflow_id__steps__step_id__branches__branch_id__delete",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Id"
            }
          },
          {
            "name": "step_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Step Id"
            }
          },
          {
            "name": "branch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Branch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStepResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoutingBranchCreate": {
        "properties": {
          "condition_result": {
            "type": "string",
            "maxLength": 100,
            "title": "Condition Result"
          },
          "action_type": {
            "type": "string",
            "maxLength": 100,
            "title": "Action Type"
          },
          "next_step_id": {
            "type": "string",
            "maxLength": 100,
            "title": "Next Step Id"
          }
        },
        "type": "object",
        "required": [
          "condition_result",
          "action_type",
          "next_step_id"
        ],
        "title": "RoutingBranchCreate",
        "description": "Schema for creating a RoutingBranch."
      },
...
      "WorkflowCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "cover_image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Url"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "deployed"
            ],
            "title": "Status",
            "default": "draft"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "WorkflowCreate",
        "description": "Schema for creating a Workflow."
      },
      "WorkflowListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkflowSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "WorkflowListResponse",
        "description": "Schema for paginated workflow list response."
      },
      "WorkflowResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "cover_image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Url"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "deployed"
            ],
            "title": "Status",
            "default": "draft"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/WorkflowStepResponse"
            },
            "type": "array",
            "title": "Steps",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "WorkflowResponse",
        "description": "Schema for Workflow response with steps."
      },
...
      "WorkflowStepResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "step_order": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Step Order"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "completed"
            ],
            "title": "Status",
            "default": "pending"
          },
          "context_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "image",
                  "text",
                  "voice"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Type"
          },
          "context_image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Image Url"
          },
          "context_text_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Text Content"
          },
          "context_voice_transcript": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Voice Transcript"
          },
          "context_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Description"
          },
          "extraction_keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Keywords"
          },
          "extraction_voice_transcript": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Voice Transcript"
          },
          "logic_strategy": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "rule_based",
                  "few_shot"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Logic Strategy"
          },
          "logic_rule_expression": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logic Rule Expression"
          },
          "logic_evaluation_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logic Evaluation Prompt"
          },
          "routing_default_next": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Routing Default Next"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "workflow_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workflow Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "examples": {
            "items": {
              "$ref": "#/components/schemas/ExampleResponse"
            },
            "type": "array",
            "title": "Examples",
            "default": []
          },
          "routing_branches": {
            "items": {
              "$ref": "#/components/schemas/RoutingBranchResponse"
            },
            "type": "array",
            "title": "Routing Branches",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "step_order",
          "id",
          "workflow_id",
          "created_at",
          "updated_at"
        ],
        "title": "WorkflowStepResponse",
        "description": "Schema for WorkflowStep response."
      },
...
      "WorkflowSummary": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "cover_image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Url"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "deployed"
            ],
            "title": "Status",
            "default": "draft"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "WorkflowSummary",
        "description": "Summary schema for Workflow (without nested steps)."
      },
...
    }
  }
}